vine: consistently use int64 in copy_file_to_file (#3863)
* consistently use int64 in copy_file_to_file

* use int for opening fd

* use int64 in putfile

* use int64 for batch_fs_putfile

* use int64 in various putfile

* check success < 0 instead of <= 0 for copy_file_to_file

+++ batch_job/src/batch_job.c
@@ -287,7 +287,7 @@ int batch_fs_mkdir (struct batch_queue *q, const char *path, mode_t mode, int re
 	return q->module->fs.mkdir(q, path, mode, recursive);
 }
 
-int batch_fs_putfile (struct batch_queue *q, const char *lpath, const char *rpath)
+int64_t batch_fs_putfile (struct batch_queue *q, const char *lpath, const char *rpath)
 {
 	return q->module->fs.putfile(q, lpath, rpath);
 }
